pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "legal-ai"
version = "0.1.0"
description = "Legal AI Assistant for Indian Legal Documents"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "fastapi",
    "uvicorn",
    "python-multipart",
    "chromadb",
    "langchain",
    "langchain-community",
    "langchain-core",
    "langchain-openai",
    "pydantic",
    "pydantic-settings",
    "python-dotenv",
    "PyPDF2",
    "pytesseract",
    "beautifulsoup4",
    "requests",
    "tqdm"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "isort",
    "flake8",
    "mypy"
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.ruff]
line-length = 88
target-version = "py39"
select = ["E", "F", "B", "I"]
